chore(tests): move faker to concern namespace
