+++ tests/Concern/Faker.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests;
+namespace App\Tests\Concern;
 
 use Faker\Factory;
 use Faker\Generator;
