+++ tests/Concern/WithFaker.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace App\Tests\Concern;
 
-use App\Tests\Faker;
-
 trait WithFaker
 {
     protected Faker $faker;
