+++ tests/FeatureTestCase.php
@@ -176,12 +176,4 @@ class FeatureTestCase extends WebTestCase
     {
         return self::getContainer();
     }
-
-    /**
-     * Returns the faker instance.
-     */
-    protected function faker(): Faker
-    {
-        return $this->faker;
-    }
 }
